+++ src/index.ts
@@ -1,4 +1,4 @@
-export { defineBddConfig } from './config';
+export { defineBddConfig, defineBddProject } from './config';
 export { BDDInputConfig } from './config/types';
 export { createBdd } from './steps/createBdd';
 export { test } from './run/testFixtures';