+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * improvement: remove dependency on Cucumber runner, see all the details in [migration guide](), ([#136](https://github.com/vitalets/playwright-bdd/issues/136))
 * improvement: support custom parameter types in decorator steps, ([#112](https://github.com/vitalets/playwright-bdd/issues/112))
 * improvement: manual `importTestFrom` option is replaced with automatic detection and most likely can be removed from BDD configuration, ([#46](https://github.com/vitalets/playwright-bdd/issues/46))
+* improvement: introduce new function `defineBddProject()` that makes it easier to configure BDD for Playwright projects, ([#169](https://github.com/vitalets/playwright-bdd/issues/169))
 * chore: set minimal Playwright version to 1.35
 * chore: set minimal Node.js version to 18